requirement.txt
numpy==1.23.5
scikit-learn==1.2.2
fastapi[standard]
tensorflow-cpu>=2.10,<2.12
